LayoutTests/platform/mac/fast/images/exif-orientation-css-expected.txt

None of the images should be rotated. This test is only valid when run with testRunner (or with WebKitShouldRespectImageOrientation manually set to true).


Normal 
Flipped horizontally 
Rotated 180° 
Flipped vertically 

Rotated 90° CCW and flipped vertically 
Rotated 90° CCW 
Rotated 90° CW and flipped vertically 
Rotated 90° CW 

Rotated 90° CCW and flipped vertically 
Rotated 90° CCW 
Rotated 90° CW and flipped vertically 
Rotated 90° CW 

Undefined (invalid value) 
img1 size = 100px by 50px
img2 size = 100px by 50px
img3 size = 100px by 50px
img4 size = 100px by 50px
img5 size = 100px by 50px
img6 size = 100px by 50px
img7 size = 100px by 50px
img8 size = 100px by 50px
img9 size = 100px by 100px
img10 size = 100px by 100px
img11 size = 100px by 100px
img12 size = 100px by 100px
img13 size = 100px by 50px
